fix(engine): compile macos engine closure
Use portable atomic shared-pointer publication and explicit vertex layout offset conversions so AppleClang can build the complete Engine and DurinEd closure. Launch now stops at the intentionally deferred macOS crash-service boundary; affected test assertions pass but their processes still fault after global teardown.

Plan: Documentation/Plans/MacOSNativeToolchainBootstrap.md

Stage: Stage 3: Configure and compile the bounded native closure

+++ Engine/Source/Runtime/Engine/Private/SkeletalMesh/SkeletalMeshVertexFactory.cpp
@@ -51,8 +51,8 @@ namespace Durin
 		{
 			Data.TextureCoordinates[Channel] = {
 				.VertexBuffer = &TexCoords,
-				.Offset = offsetof(FStaticMeshTexcoordVertex, TexCoords)
-					+ sizeof(FVector2f) * Channel,
+				.Offset = static_cast<uint32>(offsetof(FStaticMeshTexcoordVertex, TexCoords)
+					+ sizeof(FVector2f) * Channel),
 				.Stride = static_cast<uint16>(TexCoords.GetStride()),
 				.Type = EVertexElementType::Float2,
 				.AttributeIndex = static_cast<uint8>(3 + Channel),
